Split build tasks into several files
